Small changes to memory config for debugging. Refactor retiming tests
