+++ test/spatial/tests/feature/retiming/SimpleRetimePipe.scala
@@ -1,0 +1,27 @@
+package spatial.tests.feature.retiming
+
+import argon._
+import spatial.dsl._
+import spatial.node.DelayLine
+
+@spatial class SimpleRetimePipe extends SpatialTest {
+  override def backends = super.backends.filterNot{be => (be == Scala) | (be == VCS_noretime)}
+
+  def main(args: Array[String]): Unit = {
+    val a = ArgIn[Int]
+    val b = ArgIn[Int]
+    val c = ArgIn[Int]
+    val d = ArgOut[Int]
+    Accel {
+      d := a * b + c
+    }
+    println("d: " + getArg(d))
+    assert(getArg(d) == 0.to[Int])
+  }
+
+  override def checkIR(block: Block[_]): Result = {
+    val delays = block.nestedStms.count{case Op(_:DelayLine[_]) => true; case _ => false }
+    delays shouldBe 3
+    super.checkIR(block)
+  }
+}
